RAH/requirements.txt

asgiref==3.7.1

Django==3.2.19

django-rest-passwordreset==1.3.0

djangorestframework==3.14.0

Pillow==9.5.0

psycopg2==2.9.6

pytz==2023.3

sqlparse==0.4.4

typing-extensions==4.6.1